+++ bed_reader/tests/test_open_bed.py
@@ -1453,7 +1453,6 @@ def test_stream_modes2(shared_datadir, tmp_path):
                     assert np.allclose(val, val3, equal_nan=True)
 
 
-# cmk write via to_bed with both modes (or documentation or error message)
 # cmk get all tests to pass and get CI to pass
 
 if __name__ == "__main__":